Changed Class Names, Removed Private Fields, and Merged Files
