+++ index.js
@@ -42,8 +42,8 @@ export {
 	initShimAsyncCodec,
 	ZipReader,
 	ZipWriter,
-	ZipDecompressionStream,
-	ZipCompressionStream,
+	ZipReaderStream,
+	ZipWriterStream,
 	Reader,
 	Writer,
 	TextReader,
