+++ lib/core/zip-reader.js
@@ -7,8 +7,8 @@
  1. Redistributions of source code must retain the above copyright notice,
  this list of conditions and the following disclaimer.
 
- 2. Redistributions in binary form must reproduce the above copyright 
- notice, this list of conditions and the following disclaimer in 
+ 2. Redistributions in binary form must reproduce the above copyright
+ notice, this list of conditions and the following disclaimer in
  the documentation and/or other materials provided with the distribution.
 
  3. The names of the authors may not be used to endorse or promote products
@@ -209,7 +209,7 @@ class ZipReader {
 		}
 		if (directoryDataOffset >= reader.size) {
 			prependedDataLength = reader.size - directoryDataOffset - directoryDataLength - END_OF_CENTRAL_DIR_LENGTH;
-			directoryDataOffset = reader.size - directoryDataLength - END_OF_CENTRAL_DIR_LENGTH;	
+			directoryDataOffset = reader.size - directoryDataLength - END_OF_CENTRAL_DIR_LENGTH;
 		}
 		if (expectedLastDiskNumber != lastDiskNumber) {
 			throw new Error(ERR_SPLIT_ZIP_FILE);
@@ -326,8 +326,38 @@ class ZipReader {
 	}
 }
 
+class ZipReaderStream {
+	readable;
+	writable;
+	constructor (options = {}) {
+		const { readable, writable } = new TransformStream();
+		const gen = new ZipReader(readable, options).getEntriesGenerator();
+		this.readable = new ReadableStream({
+			async pull(controller) {
+				const { done, value } = await gen.next();
+				if (done)
+					return controller.close();
+				const chunk = {
+					...value,
+					readable: (function () {
+						const { readable, writable } = new TransformStream();
+						if (value.getData) {
+							value.getData(writable);
+							return readable;
+						}
+					})()
+				};
+				delete chunk.getData;
+				controller.enqueue(chunk);
+			}
+		});
+		this.writable = writable;
+	}
+}
+
 export {
 	ZipReader,
+	ZipReaderStream,
 	ERR_BAD_FORMAT,
 	ERR_EOCDR_NOT_FOUND,
 	ERR_EOCDR_ZIP64_NOT_FOUND,
