+++ lib/core/zip-writer.js
@@ -7,8 +7,8 @@
  1. Redistributions of source code must retain the above copyright notice,
  this list of conditions and the following disclaimer.
 
- 2. Redistributions in binary form must reproduce the above copyright 
- notice, this list of conditions and the following disclaimer in 
+ 2. Redistributions in binary form must reproduce the above copyright
+ notice, this list of conditions and the following disclaimer in
  the documentation and/or other materials provided with the distribution.
 
  3. The names of the authors may not be used to endorse or promote products
@@ -176,8 +176,37 @@ class ZipWriter {
 	}
 }
 
+class ZipWriterStream {
+	readable;
+	zipWriter;
+	constructor (options = {}) {
+		const { readable, writable } = new TransformStream();
+		this.readable = readable;
+		this.zipWriter = new ZipWriter(writable, options);
+	}
+
+	transform(path) {
+		const { readable, writable } = new TransformStream({
+			flush: () => { this.zipWriter.close(); }
+		});
+		this.zipWriter.add(path, readable);
+		return { readable: this.readable, writable };
+	}
+
+	writable(path) {
+		const { readable, writable } = new TransformStream();
+		this.zipWriter.add(path, readable);
+		return writable;
+	}
+
+	close(comment = undefined, options = {}) {
+		return this.zipWriter.close(comment, options);
+	}
+}
+
 export {
 	ZipWriter,
+	ZipWriterStream,
 	ERR_DUPLICATED_NAME,
 	ERR_INVALID_COMMENT,
 	ERR_INVALID_ENTRY_NAME,
