+++ lib/zip-fs.js
@@ -44,8 +44,6 @@ configureWebWorker(configure);
 export * from "./core/io.js";
 export * from "./core/zip-reader.js";
 export * from "./core/zip-writer.js";
-export * from "./core/zip-decompress.js";
-export * from "./core/zip-compress.js";
 export * from "./core/zip-fs-core.js";
 export {
 	configure,
